Move viewer model to components directory
