--- napari/components/_viewer/__init__.py
@@ -1,1 +1,0 @@
-from .viewer_model import ViewerModel
